test: move test code
